requirements.txt
ultralytics
opencv-python
numpy
scikit-learn
matplotlib
scipy
scikit-image
Pillow
torch
torchvision
tqdm
seaborn
PyWavelets